+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 * `DatabaseConfig#connect` now is deprecated. It is recommended to use `DatabaseConfig#database` property
 * Fixed issue with `PostsSchedulesTable`, when table was not created automatically
+* Refactor work with posts and posts messages
 
 ### 1.8.1